removing cyclic import
